+++ src/mathematics/square.c
@@ -2,10 +2,10 @@
 
 void square()
 {
-    float input;
+    long double input;
 
     printf("\nEnter number to square: ");
-    scanf("%f", &input);
+    scanf("%Lf", &input);
 
-    printf("\nResult: %f\n", input * input);
+    printf("\nResult: %.2Lf\n", input * input);
 }
