+++ src/mathematics/square_perimeter.c
@@ -2,10 +2,10 @@
 
 void squarePerimeter()
 {
-    float side;
+    long double side;
 
     printf("\nEnter the length of one side: ");
-    scanf("%f", &side);
+    scanf("%Lf", &side);
 
-    printf("\nResult: %f\n", side * 4);
+    printf("\nResult: %.2Lf\n", side * 4);
 }
